Force 'move' policy for closure blocks
For all capture rules to work correctly, the user must provide a closure
for the MOVE policy. Previously, we used to promote non-move closures
to move closures within macro expansions, but we felt it would be
confusing for non-move closures to implicitly move capture values,
so we've changed this to require all closures to specify the move
keyword.